requirements.txt
streamlit>=1.39.0
google-genai>=1.39.1
pandas>=2.2.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9
seaborn>=0.13.2
matplotlib>=3.8.4
langfuse>=2.0.0
python-dotenv>=1.0.1
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0
